Pulled Trainee out of utils, improved admin list view for houses and permissions groups
