+++ ap/houses/models.py
@@ -1,5 +1,6 @@
 from django.db import models
 from aputils.models import Address
+from aputils.utils import sorted_user_list_str
 
 """ HOUSES models.py
 
@@ -31,7 +32,10 @@ class House(models.Model):
 
     # whether this house is actively used by the training
     used = models.BooleanField(default=True)
-    
+
+    def residents_list(self):
+        return sorted_user_list_str(self.residents.all())
+
     #returns a query set of the empty bunks for this house
     def empty_bunk_count(self,position_list=[]):
         if len(position_list)==0:
@@ -82,7 +86,7 @@ class Bunk(models.Model):
         ('B', 'Bottom'),
         ('T', 'Top'),
         ('L', 'Queen-Left'),  # for couples
-        ('R', 'Queen-Right'), 
+        ('R', 'Queen-Right'),
         ('S', 'Single')
     )
 
@@ -112,23 +116,23 @@ class Bunk(models.Model):
 
     # which room this bunk is in
     room = models.ForeignKey(Room)
-    
+
     length = models.CharField(max_length=1, choices=LENGTH, default='R')
-    
+
     # type of bed frame
     frame = models.CharField(max_length=2, choices=FRAME_TYPES, null=True, blank=True)
-    
+
     # type of mattress
     mattress = models.CharField(max_length=50, null=True, blank=True)
 
     # whether bunk has a guardrail
     guardrail = models.NullBooleanField(blank=True)
-    
+
     # whether bunk has a ladder
     ladder = models.NullBooleanField(blank=True)
-    
+
     notes = models.TextField(blank=True, null=True)
 
     def __unicode__(self):
         return self.room.house.name + " Bunk " + str(self.number)
-    
+
